base test

+++ load.py
@@ -1,4 +1,4 @@
-"""Artemis Scanner Tracker v0.3.0 by Balvald."""
+"""Artemis Scanner Tracker v0.3.2 dev by Balvald."""
 
 import json
 import logging
@@ -6,8 +6,13 @@ import os
 import tkinter as tk
 from typing import Optional
 
-import myNotebook as nb  # type: ignore
-from config import appname  # type: ignore
+try:
+    import myNotebook as nb  # type: ignore
+    from config import appname  # type: ignore
+    testmode = False
+except ImportError:
+    import tkinter.ttk as nb  # type: ignore
+    testmode = True
 
 import eventhandling
 import ui
@@ -42,24 +47,24 @@ directory, filename = os.path.split(os.path.realpath(__file__))
 
 filenames = ["/soldbiodata.json", "/notsoldbiodata.json",  "/cmdrstates.json"]
 
-
-for file in filenames:
-    if not os.path.exists(directory + file):
-        f = open(directory + file, "w", encoding="utf8")
-        f.write(r"{}")
-        f.close()
-    elif file == "/soldbiodata.json" or file == "/notsoldbiodata.json":
-        # (not)soldbiodata file already exists
-        with open(directory + file, "r+", encoding="utf8") as f:
-            test = json.load(f)
-            if type([]) == type(test):  # noqa E721
-                # we have an old version of the (not)soldbiodata.json
-                # clear it, have the user do the journal crawling again.
-                logger.warning(f"Found old {file} format")
-                logger.warning("Clearing file...")
-                f.seek(0)
-                f.write(r"{}")
-                f.truncate()
+if not testmode:
+    for file in filenames:
+        if not os.path.exists(directory + file):
+            f = open(directory + file, "w", encoding="utf8")
+            f.write(r"{}")
+            f.close()
+        elif file == "/soldbiodata.json" or file == "/notsoldbiodata.json":
+            # (not)soldbiodata file already exists
+            with open(directory + file, "r+", encoding="utf8") as f:
+                test = json.load(f)
+                if type([]) == type(test):  # noqa E721
+                    # we have an old version of the (not)soldbiodata.json
+                    # clear it, have the user do the journal crawling again.
+                    logger.warning(f"Found old {file} format")
+                    logger.warning("Clearing file...")
+                    f.seek(0)
+                    f.write(r"{}")
+                    f.truncate()
 
 
 # load notyetsolddata and cmdrstates
